--- schema/json/binding_request_params.schema.json
@@ -1,17 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-05/schema#",
-  "title": "Binding-request-parameters-schema",
-  "description": "Schema for parameters for a service-binding creation with an instances of Application Autoscaler.",
-  "type": "object",
-
-  "properties": {
-    "configuration": {
-      "$ref": "./binding-configuration.json"
-    },
-    "scaling-policy": {
-      "$ref": "./autoscaler-policy.json"
-    }
-  },
-
-  "additionalProperties": false
-}
