+++ src/autoscaler/api/binding-requests/autoscaler-policy.json
@@ -1,0 +1,1 @@
+../../../../schema/json/autoscaler-policy.json
