+++ src/autoscaler/api/binding-requests/binding-configuration.json
@@ -1,0 +1,1 @@
+../../../../schema/json/binding-configuration.json
