+++ src/autoscaler/api/binding-requests/binding-request.json
@@ -1,0 +1,1 @@
+../../../../schema/json/binding-request.json
