+++ src/autoscaler/api/binding-requests/shared_definitions.json
@@ -1,0 +1,1 @@
+../../../../schema/json/shared_definitions.json
