--- src/autoscaler/api/binding_request_params/binding_request_params.schema.json
@@ -1,1 +1,0 @@
-../../../../schema/json/binding_request_params.schema.json
