--- src/autoscaler/api/binding_request_params/shared_definitions.json
@@ -1,1 +1,0 @@
-../../../../schema/json/shared_definitions.json
