--- src/autoscaler/api/test_shared_def.local.1up.json
@@ -1,1 +1,0 @@
-./binding_request_params/test_shared_def.local.json
